LayoutTests/inspector/model/auditTestGroup-expected.txt

Testing the functions of WI.AuditTestGroup.


== Running test suite: AuditTestGroup
-- Running test case: AuditTestGroup.fromPayload.nullObject
null

-- Running test case: AuditTestGroup.fromPayload.nonObject
null

-- Running test case: AuditTestGroup.fromPayload.emptyObject
null

-- Running test case: AuditTestGroup.fromPayload.invalidTopLevelMembers
null

-- Running test case: AuditTestGroup.fromPayload.missingSubMembers
null

-- Running test case: AuditTestGroup.fromPayload.invalidSubMembers
null

-- Running test case: AuditTestGroup.fromPayload.valid
{
  "type": "test-group",
  "name": "valid group name",
  "tests": [
    {
      "type": "test-case",
      "name": "valid test name",
      "test": "valid test function"
    }
  ]
}

-- Running test case: AuditTestGroup.fromPayload.validWithInvalidOptionals
WARN: Audit Warning: "validWithInvalidOptionals test name" has a non-string "description" value
WARN: Audit Warning: "validWithInvalidOptionals test name" has a non-string "setup" value
WARN: Audit Warning: "validWithInvalidOptionals test name" is too new to run in this Web Inspector
WARN: Audit Warning: "validWithInvalidOptionals group name" has a non-string "description" value
WARN: Audit Warning: "validWithInvalidOptionals group name" has a non-string "setup" value
WARN: Audit Warning: "validWithInvalidOptionals group name" is too new to run in this Web Inspector
{
  "type": "test-group",
  "name": "validWithInvalidOptionals group name",
  "supports": 4,
  "tests": [
    {
      "type": "test-case",
      "name": "validWithInvalidOptionals test name",
      "supports": 5,
      "test": "validWithInvalidOptionals test function"
    }
  ]
}

-- Running test case: AuditTestGroup.fromPayload.validWithValidOptionals
{
  "type": "test-group",
  "name": "validWithValidOptionals group name",
  "description": "validWithValidOptionals group description",
  "supports": 2,
  "setup": "validWithValidOptionals group setup",
  "tests": [
    {
      "type": "test-case",
      "name": "validWithValidOptionals test name",
      "description": "validWithValidOptionals test description",
      "supports": 1,
      "setup": "validWithValidOptionals test setup",
      "test": "validWithValidOptionals test function"
    }
  ]
}

-- Running test case: AuditTestGroup.fromPayload.validNested
{
  "type": "test-group",
  "name": "validNested group name",
  "description": "validNested group description",
  "supports": 2,
  "setup": "validNested group setup",
  "tests": [
    {
      "type": "test-group",
      "name": "validNested nested group name",
      "description": "validNested nested group description",
      "supports": 1,
      "setup": "validNested nested group setup",
      "tests": [
        {
          "type": "test-case",
          "name": "validNested nested test name",
          "description": "validNested nested test description",
          "supports": 0,
          "setup": "validNested nested test setup",
          "test": "validNested nested test function"
        }
      ]
    },
    {
      "type": "test-case",
      "name": "validNested test name",
      "description": "validNested test description",
      "supports": -1,
      "setup": "validNested test setup",
      "test": "validNested test function"
    }
  ]
}

-- Running test case: AuditTestGroup.Disabled.Constructor.Enabled
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

-- Running test case: AuditTestGroup.Disabled.Constructor.Disabled
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

-- Running test case: AuditTestGroup.Disabled.ChangeGroup.Enabled
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

Marking group as disabled...
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

-- Running test case: AuditTestGroup.Disabled.ChangeGroup.DisabledGroup
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

Marking group as enabled...
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

-- Running test case: AuditTestGroup.Disabled.ChangeGroup.DisabledTests
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

Marking group as enabled...
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

-- Running test case: AuditTestGroup.Disabled.ChangeTests.Enabled
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

Marking Test1 as disabled...
PASS: Group should not be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should not be disabled.

Marking Test2 as disabled...
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

-- Running test case: AuditTestGroup.Disabled.ChangeTests.DisabledGroup
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

Marking test as enabled 1...
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should be disabled.

Marking test as enabled 2...
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

-- Running test case: AuditTestGroup.Disabled.ChangeTests.DisabledTests
PASS: Group should be disabled.
PASS: Test1 should be disabled.
PASS: Test2 should be disabled.

Marking test as enabled 1...
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should be disabled.

Marking test as enabled 2...
PASS: Group should not be disabled.
PASS: Test1 should not be disabled.
PASS: Test2 should not be disabled.

-- Running test case: AuditTestGroup.Supports.Constructor.Supported
PASS: Group should be supported.
PASS: Test1 should be supported.
PASS: Test2 should be supported.

-- Running test case: AuditTestGroup.Supports.Constructor.Unsupported
WARN: Audit Warning: "Group" is too new to run in this Web Inspector
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

-- Running test case: AuditTestGroup.Supports.ChangeGroup.Supported
PASS: Group should be supported.
PASS: Test1 should be supported.
PASS: Test2 should be supported.

Marking group as unsupported...
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

-- Running test case: AuditTestGroup.Supports.ChangeGroup.UnsupportedGroup
WARN: Audit Warning: "Group" is too new to run in this Web Inspector
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

Marking group as supported...
PASS: Group should be supported.
PASS: Test1 should be supported.
PASS: Test2 should be supported.

-- Running test case: AuditTestGroup.Supports.ChangeGroup.UnsupportedTests
WARN: Audit Warning: "Test1" is too new to run in this Web Inspector
WARN: Audit Warning: "Test2" is too new to run in this Web Inspector
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

Marking group as supported...
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

-- Running test case: AuditTestGroup.Supports.ChangeTests.Supported
PASS: Group should be supported.
PASS: Test1 should be supported.
PASS: Test2 should be supported.

Marking Test1 as unsupported...
PASS: Group should be supported.
PASS: Test1 should not be supported.
PASS: Test2 should be supported.

Marking Test2 as unsupported...
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

-- Running test case: AuditTestGroup.Supports.ChangeTests.UnsupportedGroup
WARN: Audit Warning: "Group" is too new to run in this Web Inspector
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

Marking Test1 as supported...
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

Marking Test2 as supported...
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

-- Running test case: AuditTestGroup.Supports.ChangeTests.UnsupportedTests
WARN: Audit Warning: "Test1" is too new to run in this Web Inspector
WARN: Audit Warning: "Test2" is too new to run in this Web Inspector
PASS: Group should not be supported.
PASS: Test1 should not be supported.
PASS: Test2 should not be supported.

Marking Test1 as supported...
PASS: Group should be supported.
PASS: Test1 should be supported.
PASS: Test2 should not be supported.

Marking Test2 as supported...
PASS: Group should be supported.
PASS: Test1 should be supported.
PASS: Test2 should be supported.
